Split out recommended themes action creators
