+++ client/state/themes/actions/index.js
@@ -1,17 +1,3 @@
-/**
- * Internal dependencies
- */
-import wpcom from 'lib/wp';
-import {
-	RECOMMENDED_THEMES_FAIL,
-	RECOMMENDED_THEMES_FETCH,
-	RECOMMENDED_THEMES_SUCCESS,
-} from 'state/action-types';
-
-import 'state/data-layer/wpcom/theme-filters';
-
-import 'state/themes/init';
-
 export { setBackPath } from 'state/themes/actions/set-back-path';
 export { receiveThemes } from 'state/themes/actions/receive-themes';
 export { receiveTheme } from 'state/themes/actions/receive-theme';
@@ -43,40 +29,7 @@ export { hideAutoLoadingHomepageWarning } from 'state/themes/actions/hide-auto-l
 export { acceptAutoLoadingHomepageWarning } from 'state/themes/actions/accept-auto-loading-homepage-warning';
 export { hideThemePreview } from 'state/themes/actions/hide-theme-preview';
 export { requestThemeFilters } from 'state/themes/actions/request-theme-filters';
-
-/**
- * Receives themes and dispatches them with recommended themes success signal.
- *
- * @param {Array} themes array of received theme objects
- * @returns {Function} Action thunk
- */
-export function receiveRecommendedThemes( themes ) {
-	return dispatch => {
-		dispatch( { type: RECOMMENDED_THEMES_SUCCESS, payload: themes } );
-	};
-}
-
-/**
- * Initiates network request for recommended themes.
- * Recommended themes are template first themes and are denoted by the 'auto-loading-homepage' tag.
- *
- * @returns {Function} Action thunk
- */
-export function getRecommendedThemes() {
-	return async dispatch => {
-		dispatch( { type: RECOMMENDED_THEMES_FETCH } );
-		const query = {
-			search: '',
-			number: 50,
-			tier: '',
-			filter: 'auto-loading-homepage',
-			apiVersion: '1.2',
-		};
-		try {
-			const res = await wpcom.undocumented().themes( null, query );
-			dispatch( receiveRecommendedThemes( res ) );
-		} catch ( error ) {
-			dispatch( { type: RECOMMENDED_THEMES_FAIL } );
-		}
-	};
-}
+export {
+	getRecommendedThemes,
+	receiveRecommendedThemes,
+} from 'state/themes/actions/recommended-themes';
